+++ src/resources/Identity/index.ts
@@ -28,12 +28,12 @@ export class Identity {
 export declare namespace Identity {
   export import VerificationReportListParams = IdentityNamespace0.VerificationReportListParams;
   export import VerificationReportRetrieveParams = IdentityNamespace0.VerificationReportRetrieveParams;
-  export {VerificationReport};
+  export {VerificationReport, VerificationReportResource};
   export import VerificationSessionListParams = IdentityNamespace1.VerificationSessionListParams;
   export import VerificationSessionCreateParams = IdentityNamespace1.VerificationSessionCreateParams;
   export import VerificationSessionRetrieveParams = IdentityNamespace1.VerificationSessionRetrieveParams;
   export import VerificationSessionUpdateParams = IdentityNamespace1.VerificationSessionUpdateParams;
   export import VerificationSessionCancelParams = IdentityNamespace1.VerificationSessionCancelParams;
   export import VerificationSessionRedactParams = IdentityNamespace1.VerificationSessionRedactParams;
-  export {VerificationSession};
+  export {VerificationSession, VerificationSessionResource};
 }