+++ src/resources/V2/Core/index.ts
@@ -65,9 +65,9 @@ export declare namespace Core {
   export import AccountRetrieveParams = V2Namespace0.Core.AccountRetrieveParams;
   export import AccountUpdateParams = V2Namespace0.Core.AccountUpdateParams;
   export import AccountCloseParams = V2Namespace0.Core.AccountCloseParams;
-  export {Account};
+  export {Account, AccountResource};
   export import AccountLinkCreateParams = V2Namespace1.Core.AccountLinkCreateParams;
-  export {AccountLink};
+  export {AccountLink, AccountLinkResource};
   export import AccountTokenCreateParams = V2Namespace2.Core.AccountTokenCreateParams;
   export import AccountTokenRetrieveParams = V2Namespace2.Core.AccountTokenRetrieveParams;
   export {AccountToken};