+++ src/stripe.esm.node.ts
@@ -81,6 +81,9 @@ import {
   AccountRetrieveCapabilityParams,
   AccountRetrieveExternalAccountParams,
   AccountRetrievePersonParams,
+  AccountRetrieveSignalsParams,
+  AccountSerializeBatchCreateParams,
+  AccountSerializeBatchDeleteParams,
   AccountSerializeBatchUpdateParams,
   AccountUpdateCapabilityParams,
   AccountUpdateExternalAccountParams,
@@ -141,6 +144,11 @@ import {
   BalanceTransactionResource,
 } from './resources/BalanceTransactions.js';
 import {
+  BalanceTransfer,
+  BalanceTransferCreateParams,
+  BalanceTransferResource,
+} from './resources/BalanceTransfers.js';
+import {
   Charge,
   ChargeCreateParams,
   ChargeRetrieveParams,
@@ -170,6 +178,8 @@ import {
   CouponListParams,
   CouponDeleteParams,
   CouponSerializeBatchCreateParams,
+  CouponSerializeBatchDeleteParams,
+  CouponSerializeBatchUpdateParams,
   CouponResource,
 } from './resources/Coupons.js';
 import {
@@ -181,6 +191,7 @@ import {
   CreditNoteListLineItemsParams,
   CreditNoteListPreviewLineItemsParams,
   CreditNotePreviewParams,
+  CreditNoteSerializeBatchCreateParams,
   CreditNoteVoidCreditNoteParams,
   CreditNoteResource,
 } from './resources/CreditNotes.js';
@@ -211,7 +222,20 @@ import {
   CustomerRetrieveSourceParams,
   CustomerRetrieveTaxIdParams,
   CustomerSearchParams,
+  CustomerSerializeBatchCreateParams,
+  CustomerSerializeBatchCreateBalanceTransactionParams,
+  CustomerSerializeBatchCreateForCustomerTaxIdParams,
+  CustomerSerializeBatchCreateFundingInstructionsFundingInstructionsParams,
+  CustomerSerializeBatchCreateSourceParams,
+  CustomerSerializeBatchDeleteParams,
+  CustomerSerializeBatchDeleteDiscountParams,
+  CustomerSerializeBatchDeleteSourceParams,
+  CustomerSerializeBatchDeleteTaxIdParams,
   CustomerSerializeBatchUpdateParams,
+  CustomerSerializeBatchUpdateBalanceTransactionParams,
+  CustomerSerializeBatchUpdateCashBalanceParams,
+  CustomerSerializeBatchUpdateSourceParams,
+  CustomerSerializeBatchVerifySourceParams,
   CustomerUpdateBalanceTransactionParams,
   CustomerUpdateCashBalanceParams,
   CustomerUpdateSourceParams,
@@ -221,6 +245,7 @@ import {
 import {
   CustomerSession,
   CustomerSessionCreateParams,
+  CustomerSessionSerializeBatchCreateParams,
   CustomerSessionResource,
 } from './resources/CustomerSessions.js';
 import {
@@ -229,6 +254,7 @@ import {
   DisputeUpdateParams,
   DisputeListParams,
   DisputeCloseParams,
+  DisputeSerializeBatchCloseParams,
   DisputeResource,
 } from './resources/Disputes.js';
 import {
@@ -269,6 +295,14 @@ import {
   FileLinkResource,
 } from './resources/FileLinks.js';
 import {
+  FrMealVouchersOnboarding,
+  FrMealVouchersOnboardingCreateParams,
+  FrMealVouchersOnboardingRetrieveParams,
+  FrMealVouchersOnboardingUpdateParams,
+  FrMealVouchersOnboardingListParams,
+  FrMealVouchersOnboardingResource,
+} from './resources/FrMealVouchersOnboardings.js';
+import {
   FxQuote,
   FxQuoteCreateParams,
   FxQuoteRetrieveParams,
@@ -294,6 +328,19 @@ import {
   InvoiceRemoveLinesParams,
   InvoiceSearchParams,
   InvoiceSendInvoiceParams,
+  InvoiceSerializeBatchAddLinesParams,
+  InvoiceSerializeBatchCreateParams,
+  InvoiceSerializeBatchCreatePreviewParams,
+  InvoiceSerializeBatchDeleteParams,
+  InvoiceSerializeBatchFinalizeInvoiceParams,
+  InvoiceSerializeBatchMarkUncollectibleParams,
+  InvoiceSerializeBatchPayParams,
+  InvoiceSerializeBatchRemoveLinesParams,
+  InvoiceSerializeBatchSendInvoiceParams,
+  InvoiceSerializeBatchUpdateParams,
+  InvoiceSerializeBatchUpdateLineItemParams,
+  InvoiceSerializeBatchUpdateLinesParams,
+  InvoiceSerializeBatchVoidInvoiceParams,
   InvoiceUpdateLinesParams,
   InvoiceUpdateLineItemParams,
   InvoiceVoidInvoiceParams,
@@ -307,6 +354,9 @@ import {
   InvoiceItemUpdateParams,
   InvoiceItemListParams,
   InvoiceItemDeleteParams,
+  InvoiceItemSerializeBatchCreateParams,
+  InvoiceItemSerializeBatchDeleteParams,
+  InvoiceItemSerializeBatchUpdateParams,
   InvoiceItemResource,
 } from './resources/InvoiceItems.js';
 import {
@@ -320,6 +370,8 @@ import {
   InvoiceRenderingTemplateRetrieveParams,
   InvoiceRenderingTemplateListParams,
   InvoiceRenderingTemplateArchiveParams,
+  InvoiceRenderingTemplateSerializeBatchArchiveParams,
+  InvoiceRenderingTemplateSerializeBatchUnarchiveParams,
   InvoiceRenderingTemplateUnarchiveParams,
   InvoiceRenderingTemplateResource,
 } from './resources/InvoiceRenderingTemplates.js';
@@ -350,6 +402,13 @@ import {
   PaymentAttemptRecord,
   PaymentAttemptRecordRetrieveParams,
   PaymentAttemptRecordListParams,
+  PaymentAttemptRecordReportAuthenticatedParams,
+  PaymentAttemptRecordReportAuthorizedParams,
+  PaymentAttemptRecordReportCanceledParams,
+  PaymentAttemptRecordReportFailedParams,
+  PaymentAttemptRecordReportGuaranteedParams,
+  PaymentAttemptRecordReportInformationalParams,
+  PaymentAttemptRecordReportRefundParams,
   PaymentAttemptRecordResource,
 } from './resources/PaymentAttemptRecords.js';
 import {
@@ -365,6 +424,7 @@ import {
   PaymentIntentDecrementAuthorizationParams,
   PaymentIntentIncrementAuthorizationParams,
   PaymentIntentListAmountDetailsLineItemsParams,
+  PaymentIntentReauthorizeParams,
   PaymentIntentSearchParams,
   PaymentIntentTriggerActionParams,
   PaymentIntentVerifyMicrodepositsParams,
@@ -380,13 +440,32 @@ import {
   PaymentLinkResource,
 } from './resources/PaymentLinks.js';
 import {
+  PaymentLocation,
+  DeletedPaymentLocation,
+  PaymentLocationCreateParams,
+  PaymentLocationRetrieveParams,
+  PaymentLocationUpdateParams,
+  PaymentLocationListParams,
+  PaymentLocationDeleteParams,
+  PaymentLocationResource,
+} from './resources/PaymentLocations.js';
+import {
+  PaymentLocationCapability,
+  PaymentLocationCapabilityRetrieveParams,
+  PaymentLocationCapabilityUpdateParams,
+  PaymentLocationCapabilityListParams,
+  PaymentLocationCapabilityResource,
+} from './resources/PaymentLocationCapabilities.js';
+import {
   PaymentMethod,
   PaymentMethodCreateParams,
   PaymentMethodRetrieveParams,
   PaymentMethodUpdateParams,
   PaymentMethodListParams,
   PaymentMethodAttachParams,
+  PaymentMethodCheckBalanceParams,
   PaymentMethodDetachParams,
+  PaymentMethodSerializeBatchAttachParams,
   PaymentMethodResource,
 } from './resources/PaymentMethods.js';
 import {
@@ -446,6 +525,8 @@ import {
   PriceUpdateParams,
   PriceListParams,
   PriceSearchParams,
+  PriceSerializeBatchCreateParams,
+  PriceSerializeBatchUpdateParams,
   PriceResource,
 } from './resources/Prices.js';
 import {
@@ -461,6 +542,11 @@ import {
   ProductListFeaturesParams,
   ProductRetrieveFeatureParams,
   ProductSearchParams,
+  ProductSerializeBatchCreateParams,
+  ProductSerializeBatchCreateFeatureParams,
+  ProductSerializeBatchDeleteParams,
+  ProductSerializeBatchDeleteFeatureParams,
+  ProductSerializeBatchUpdateParams,
   ProductResource,
 } from './resources/Products.js';
 import {
@@ -501,6 +587,8 @@ import {
   RefundUpdateParams,
   RefundListParams,
   RefundCancelParams,
+  RefundSerializeBatchCancelParams,
+  RefundSerializeBatchCreateParams,
   RefundResource,
 } from './resources/Refunds.js';
 import {
@@ -549,6 +637,7 @@ import {
   SubscriptionRetrieveParams,
   SubscriptionUpdateParams,
   SubscriptionListParams,
+  SubscriptionAttachCadenceParams,
   SubscriptionCancelParams,
   SubscriptionDeleteDiscountParams,
   SubscriptionMigrateParams,
@@ -556,7 +645,10 @@ import {
   SubscriptionResumeParams,
   SubscriptionSearchParams,
   SubscriptionSerializeBatchCancelParams,
+  SubscriptionSerializeBatchCreateParams,
   SubscriptionSerializeBatchMigrateParams,
+  SubscriptionSerializeBatchPauseParams,
+  SubscriptionSerializeBatchResumeParams,
   SubscriptionSerializeBatchUpdateParams,
   SubscriptionResource,
 } from './resources/Subscriptions.js';
@@ -568,6 +660,9 @@ import {
   SubscriptionItemUpdateParams,
   SubscriptionItemListParams,
   SubscriptionItemDeleteParams,
+  SubscriptionItemSerializeBatchCreateParams,
+  SubscriptionItemSerializeBatchDeleteParams,
+  SubscriptionItemSerializeBatchUpdateParams,
   SubscriptionItemResource,
 } from './resources/SubscriptionItems.js';
 import {
@@ -579,6 +674,10 @@ import {
   SubscriptionScheduleAmendParams,
   SubscriptionScheduleCancelParams,
   SubscriptionScheduleReleaseParams,
+  SubscriptionScheduleSerializeBatchCancelParams,
+  SubscriptionScheduleSerializeBatchCreateParams,
+  SubscriptionScheduleSerializeBatchReleaseParams,
+  SubscriptionScheduleSerializeBatchUpdateParams,
   SubscriptionScheduleResource,
 } from './resources/SubscriptionSchedules.js';
 import {
@@ -594,6 +693,7 @@ import {
   TaxIdRetrieveParams,
   TaxIdListParams,
   TaxIdDeleteParams,
+  TaxIdSerializeBatchCreateParams,
   TaxIdResource,
 } from './resources/TaxIds.js';
 import {
@@ -602,6 +702,8 @@ import {
   TaxRateRetrieveParams,
   TaxRateUpdateParams,
   TaxRateListParams,
+  TaxRateSerializeBatchCreateParams,
+  TaxRateSerializeBatchUpdateParams,
   TaxRateResource,
 } from './resources/TaxRates.js';
 import {
@@ -649,11 +751,16 @@ import {ConnectCollectionTransfer} from './resources/ConnectCollectionTransfers.
 import {Discount, DeletedDiscount} from './resources/Discounts.js';
 import {FundingInstructions} from './resources/FundingInstructions.js';
 import {LineItem} from './resources/LineItems.js';
+import {PaymentMethodBalance} from './resources/PaymentMethodBalances.js';
+import {Profile} from './resources/Profiles.js';
 import {QuoteLine} from './resources/QuoteLines.js';
 import {ReserveTransaction} from './resources/ReserveTransactions.js';
+import {RiskSignals} from './resources/RiskSignals.js';
 import {SourceMandateNotification} from './resources/SourceMandateNotifications.js';
 import {SourceTransaction} from './resources/SourceTransactions.js';
 import {TaxDeductedAtSource} from './resources/TaxDeductedAtSources.js';
+import {TransitBalance} from './resources/TransitBalances.js';
+import {AccountSignals} from './resources/AccountSignals.js';
 import {Capability} from './resources/Capabilities.js';
 import {LoginLink} from './resources/LoginLinks.js';
 import {Person, DeletedPerson} from './resources/Persons.js';
@@ -681,11 +788,13 @@ import {BillingPortal} from './resources/BillingPortal/index.js';
 import {Capital} from './resources/Capital/index.js';
 import {Checkout} from './resources/Checkout/index.js';
 import {Climate} from './resources/Climate/index.js';
+import {DelegatedCheckout} from './resources/DelegatedCheckout/index.js';
 import {Entitlements} from './resources/Entitlements/index.js';
 import {FinancialConnections} from './resources/FinancialConnections/index.js';
 import {Forwarding} from './resources/Forwarding/index.js';
 import {Identity} from './resources/Identity/index.js';
 import {Issuing} from './resources/Issuing/index.js';
+import {Orchestration} from './resources/Orchestration/index.js';
 import {Privacy} from './resources/Privacy/index.js';
 import {ProductCatalog} from './resources/ProductCatalog/index.js';
 import {Radar} from './resources/Radar/index.js';
@@ -699,6 +808,7 @@ import {TestHelpers} from './resources/TestHelpers/index.js';
 import {Treasury} from './resources/Treasury/index.js';
 import {V2} from './resources/V2/index.js';
 // StripeInstanceImports: The end of the section generated from our OpenAPI spec
+
 // V1EventImports: The beginning of the section generated from our OpenAPI spec
 import {
   AccountApplicationAuthorizedEvent,
@@ -714,6 +824,7 @@ import {
   ApplicationFeeRefundedEvent,
   BalanceAvailableEvent,
   BalanceSettingsUpdatedEvent,
+  BillingAlertRecoveredEvent,
   BillingAlertTriggeredEvent,
   BillingCreditBalanceTransactionCreatedEvent,
   BillingCreditGrantCreatedEvent,
@@ -865,6 +976,7 @@ import {
   PaymentIntentAmountCapturableUpdatedEvent,
   PaymentIntentCanceledEvent,
   PaymentIntentCreatedEvent,
+  PaymentIntentExpiredEvent,
   PaymentIntentPartiallyFundedEvent,
   PaymentIntentPaymentFailedEvent,
   PaymentIntentProcessingEvent,
@@ -1112,6 +1224,7 @@ export class Stripe {
   balance: BalanceResource;
   balanceSettings: BalanceSettingResource;
   balanceTransactions: BalanceTransactionResource;
+  balanceTransfers: BalanceTransferResource;
   charges: ChargeResource;
   confirmationTokens: ConfirmationTokenResource;
   countrySpecs: CountrySpecResource;
@@ -1126,6 +1239,7 @@ export class Stripe {
   externalAccounts: ExternalAccountResource;
   fileLinks: FileLinkResource;
   files: FileResource;
+  frMealVouchersOnboardings: FrMealVouchersOnboardingResource;
   fxQuotes: FxQuoteResource;
   invoiceItems: InvoiceItemResource;
   invoicePayments: InvoicePaymentResource;
@@ -1137,6 +1251,8 @@ export class Stripe {
   paymentAttemptRecords: PaymentAttemptRecordResource;
   paymentIntents: PaymentIntentResource;
   paymentLinks: PaymentLinkResource;
+  paymentLocationCapabilities: PaymentLocationCapabilityResource;
+  paymentLocations: PaymentLocationResource;
   paymentMethodConfigurations: PaymentMethodConfigurationResource;
   paymentMethodDomains: PaymentMethodDomainResource;
   paymentMethods: PaymentMethodResource;
@@ -1169,11 +1285,13 @@ export class Stripe {
   capital: Capital;
   checkout: Checkout;
   climate: Climate;
+  delegatedCheckout: DelegatedCheckout;
   entitlements: Entitlements;
   financialConnections: FinancialConnections;
   forwarding: Forwarding;
   identity: Identity;
   issuing: Issuing;
+  orchestration: Orchestration;
   privacy: Privacy;
   productCatalog: ProductCatalog;
   radar: Radar;
@@ -1298,6 +1416,7 @@ export class Stripe {
     this.balance = new BalanceResource(this);
     this.balanceSettings = new BalanceSettingResource(this);
     this.balanceTransactions = new BalanceTransactionResource(this);
+    this.balanceTransfers = new BalanceTransferResource(this);
     this.charges = new ChargeResource(this);
     this.confirmationTokens = new ConfirmationTokenResource(this);
     this.countrySpecs = new CountrySpecResource(this);
@@ -1312,6 +1431,7 @@ export class Stripe {
     this.externalAccounts = new ExternalAccountResource(this);
     this.fileLinks = new FileLinkResource(this);
     this.files = new FileResource(this);
+    this.frMealVouchersOnboardings = new FrMealVouchersOnboardingResource(this);
     this.fxQuotes = new FxQuoteResource(this);
     this.invoiceItems = new InvoiceItemResource(this);
     this.invoicePayments = new InvoicePaymentResource(this);
@@ -1323,6 +1443,10 @@ export class Stripe {
     this.paymentAttemptRecords = new PaymentAttemptRecordResource(this);
     this.paymentIntents = new PaymentIntentResource(this);
     this.paymentLinks = new PaymentLinkResource(this);
+    this.paymentLocationCapabilities = new PaymentLocationCapabilityResource(
+      this
+    );
+    this.paymentLocations = new PaymentLocationResource(this);
     this.paymentMethodConfigurations = new PaymentMethodConfigurationResource(
       this
     );
@@ -1357,11 +1481,13 @@ export class Stripe {
     this.capital = new Capital(this);
     this.checkout = new Checkout(this);
     this.climate = new Climate(this);
+    this.delegatedCheckout = new DelegatedCheckout(this);
     this.entitlements = new Entitlements(this);
     this.financialConnections = new FinancialConnections(this);
     this.forwarding = new Forwarding(this);
     this.identity = new Identity(this);
     this.issuing = new Issuing(this);
+    this.orchestration = new Orchestration(this);
     this.privacy = new Privacy(this);
     this.productCatalog = new ProductCatalog(this);
     this.radar = new Radar(this);
@@ -1899,6 +2025,9 @@ export declare namespace Stripe {
     AccountRetrieveCapabilityParams,
     AccountRetrieveExternalAccountParams,
     AccountRetrievePersonParams,
+    AccountRetrieveSignalsParams,
+    AccountSerializeBatchCreateParams,
+    AccountSerializeBatchDeleteParams,
     AccountSerializeBatchUpdateParams,
     AccountUpdateCapabilityParams,
     AccountUpdateExternalAccountParams,
@@ -1947,6 +2076,11 @@ export declare namespace Stripe {
     BalanceTransactionResource,
   };
   export {
+    BalanceTransfer,
+    BalanceTransferCreateParams,
+    BalanceTransferResource,
+  };
+  export {
     Charge,
     ChargeCreateParams,
     ChargeRetrieveParams,
@@ -1976,6 +2110,8 @@ export declare namespace Stripe {
     CouponListParams,
     CouponDeleteParams,
     CouponSerializeBatchCreateParams,
+    CouponSerializeBatchDeleteParams,
+    CouponSerializeBatchUpdateParams,
     CouponResource,
   };
   export {
@@ -1987,6 +2123,7 @@ export declare namespace Stripe {
     CreditNoteListLineItemsParams,
     CreditNoteListPreviewLineItemsParams,
     CreditNotePreviewParams,
+    CreditNoteSerializeBatchCreateParams,
     CreditNoteVoidCreditNoteParams,
     CreditNoteResource,
   };
@@ -2017,7 +2154,20 @@ export declare namespace Stripe {
     CustomerRetrieveSourceParams,
     CustomerRetrieveTaxIdParams,
     CustomerSearchParams,
+    CustomerSerializeBatchCreateParams,
+    CustomerSerializeBatchCreateBalanceTransactionParams,
+    CustomerSerializeBatchCreateForCustomerTaxIdParams,
+    CustomerSerializeBatchCreateFundingInstructionsFundingInstructionsParams,
+    CustomerSerializeBatchCreateSourceParams,
+    CustomerSerializeBatchDeleteParams,
+    CustomerSerializeBatchDeleteDiscountParams,
+    CustomerSerializeBatchDeleteSourceParams,
+    CustomerSerializeBatchDeleteTaxIdParams,
     CustomerSerializeBatchUpdateParams,
+    CustomerSerializeBatchUpdateBalanceTransactionParams,
+    CustomerSerializeBatchUpdateCashBalanceParams,
+    CustomerSerializeBatchUpdateSourceParams,
+    CustomerSerializeBatchVerifySourceParams,
     CustomerUpdateBalanceTransactionParams,
     CustomerUpdateCashBalanceParams,
     CustomerUpdateSourceParams,
@@ -2027,6 +2177,7 @@ export declare namespace Stripe {
   export {
     CustomerSession,
     CustomerSessionCreateParams,
+    CustomerSessionSerializeBatchCreateParams,
     CustomerSessionResource,
   };
   export {
@@ -2035,6 +2186,7 @@ export declare namespace Stripe {
     DisputeUpdateParams,
     DisputeListParams,
     DisputeCloseParams,
+    DisputeSerializeBatchCloseParams,
     DisputeResource,
   };
   export {
@@ -2082,6 +2234,14 @@ export declare namespace Stripe {
     FileLinkResource,
   };
   export {
+    FrMealVouchersOnboarding,
+    FrMealVouchersOnboardingCreateParams,
+    FrMealVouchersOnboardingRetrieveParams,
+    FrMealVouchersOnboardingUpdateParams,
+    FrMealVouchersOnboardingListParams,
+    FrMealVouchersOnboardingResource,
+  };
+  export {
     FxQuote,
     FxQuoteCreateParams,
     FxQuoteRetrieveParams,
@@ -2107,6 +2267,19 @@ export declare namespace Stripe {
     InvoiceRemoveLinesParams,
     InvoiceSearchParams,
     InvoiceSendInvoiceParams,
+    InvoiceSerializeBatchAddLinesParams,
+    InvoiceSerializeBatchCreateParams,
+    InvoiceSerializeBatchCreatePreviewParams,
+    InvoiceSerializeBatchDeleteParams,
+    InvoiceSerializeBatchFinalizeInvoiceParams,
+    InvoiceSerializeBatchMarkUncollectibleParams,
+    InvoiceSerializeBatchPayParams,
+    InvoiceSerializeBatchRemoveLinesParams,
+    InvoiceSerializeBatchSendInvoiceParams,
+    InvoiceSerializeBatchUpdateParams,
+    InvoiceSerializeBatchUpdateLineItemParams,
+    InvoiceSerializeBatchUpdateLinesParams,
+    InvoiceSerializeBatchVoidInvoiceParams,
     InvoiceUpdateLinesParams,
     InvoiceUpdateLineItemParams,
     InvoiceVoidInvoiceParams,
@@ -2120,6 +2293,9 @@ export declare namespace Stripe {
     InvoiceItemUpdateParams,
     InvoiceItemListParams,
     InvoiceItemDeleteParams,
+    InvoiceItemSerializeBatchCreateParams,
+    InvoiceItemSerializeBatchDeleteParams,
+    InvoiceItemSerializeBatchUpdateParams,
     InvoiceItemResource,
   };
   export {
@@ -2133,6 +2309,8 @@ export declare namespace Stripe {
     InvoiceRenderingTemplateRetrieveParams,
     InvoiceRenderingTemplateListParams,
     InvoiceRenderingTemplateArchiveParams,
+    InvoiceRenderingTemplateSerializeBatchArchiveParams,
+    InvoiceRenderingTemplateSerializeBatchUnarchiveParams,
     InvoiceRenderingTemplateUnarchiveParams,
     InvoiceRenderingTemplateResource,
   };
@@ -2158,6 +2336,13 @@ export declare namespace Stripe {
     PaymentAttemptRecord,
     PaymentAttemptRecordRetrieveParams,
     PaymentAttemptRecordListParams,
+    PaymentAttemptRecordReportAuthenticatedParams,
+    PaymentAttemptRecordReportAuthorizedParams,
+    PaymentAttemptRecordReportCanceledParams,
+    PaymentAttemptRecordReportFailedParams,
+    PaymentAttemptRecordReportGuaranteedParams,
+    PaymentAttemptRecordReportInformationalParams,
+    PaymentAttemptRecordReportRefundParams,
     PaymentAttemptRecordResource,
   };
   export {
@@ -2173,6 +2358,7 @@ export declare namespace Stripe {
     PaymentIntentDecrementAuthorizationParams,
     PaymentIntentIncrementAuthorizationParams,
     PaymentIntentListAmountDetailsLineItemsParams,
+    PaymentIntentReauthorizeParams,
     PaymentIntentSearchParams,
     PaymentIntentTriggerActionParams,
     PaymentIntentVerifyMicrodepositsParams,
@@ -2188,13 +2374,32 @@ export declare namespace Stripe {
     PaymentLinkResource,
   };
   export {
+    PaymentLocation,
+    DeletedPaymentLocation,
+    PaymentLocationCreateParams,
+    PaymentLocationRetrieveParams,
+    PaymentLocationUpdateParams,
+    PaymentLocationListParams,
+    PaymentLocationDeleteParams,
+    PaymentLocationResource,
+  };
+  export {
+    PaymentLocationCapability,
+    PaymentLocationCapabilityRetrieveParams,
+    PaymentLocationCapabilityUpdateParams,
+    PaymentLocationCapabilityListParams,
+    PaymentLocationCapabilityResource,
+  };
+  export {
     PaymentMethod,
     PaymentMethodCreateParams,
     PaymentMethodRetrieveParams,
     PaymentMethodUpdateParams,
     PaymentMethodListParams,
     PaymentMethodAttachParams,
+    PaymentMethodCheckBalanceParams,
     PaymentMethodDetachParams,
+    PaymentMethodSerializeBatchAttachParams,
     PaymentMethodResource,
   };
   export {
@@ -2254,6 +2459,8 @@ export declare namespace Stripe {
     PriceUpdateParams,
     PriceListParams,
     PriceSearchParams,
+    PriceSerializeBatchCreateParams,
+    PriceSerializeBatchUpdateParams,
     PriceResource,
   };
   export {
@@ -2269,6 +2476,11 @@ export declare namespace Stripe {
     ProductListFeaturesParams,
     ProductRetrieveFeatureParams,
     ProductSearchParams,
+    ProductSerializeBatchCreateParams,
+    ProductSerializeBatchCreateFeatureParams,
+    ProductSerializeBatchDeleteParams,
+    ProductSerializeBatchDeleteFeatureParams,
+    ProductSerializeBatchUpdateParams,
     ProductResource,
   };
   export {
@@ -2309,6 +2521,8 @@ export declare namespace Stripe {
     RefundUpdateParams,
     RefundListParams,
     RefundCancelParams,
+    RefundSerializeBatchCancelParams,
+    RefundSerializeBatchCreateParams,
     RefundResource,
   };
   export {
@@ -2353,6 +2567,7 @@ export declare namespace Stripe {
     SubscriptionRetrieveParams,
     SubscriptionUpdateParams,
     SubscriptionListParams,
+    SubscriptionAttachCadenceParams,
     SubscriptionCancelParams,
     SubscriptionDeleteDiscountParams,
     SubscriptionMigrateParams,
@@ -2360,7 +2575,10 @@ export declare namespace Stripe {
     SubscriptionResumeParams,
     SubscriptionSearchParams,
     SubscriptionSerializeBatchCancelParams,
+    SubscriptionSerializeBatchCreateParams,
     SubscriptionSerializeBatchMigrateParams,
+    SubscriptionSerializeBatchPauseParams,
+    SubscriptionSerializeBatchResumeParams,
     SubscriptionSerializeBatchUpdateParams,
     SubscriptionResource,
   };
@@ -2372,6 +2590,9 @@ export declare namespace Stripe {
     SubscriptionItemUpdateParams,
     SubscriptionItemListParams,
     SubscriptionItemDeleteParams,
+    SubscriptionItemSerializeBatchCreateParams,
+    SubscriptionItemSerializeBatchDeleteParams,
+    SubscriptionItemSerializeBatchUpdateParams,
     SubscriptionItemResource,
   };
   export {
@@ -2383,6 +2604,10 @@ export declare namespace Stripe {
     SubscriptionScheduleAmendParams,
     SubscriptionScheduleCancelParams,
     SubscriptionScheduleReleaseParams,
+    SubscriptionScheduleSerializeBatchCancelParams,
+    SubscriptionScheduleSerializeBatchCreateParams,
+    SubscriptionScheduleSerializeBatchReleaseParams,
+    SubscriptionScheduleSerializeBatchUpdateParams,
     SubscriptionScheduleResource,
   };
   export {TaxCode, TaxCodeRetrieveParams, TaxCodeListParams, TaxCodeResource};
@@ -2393,6 +2618,7 @@ export declare namespace Stripe {
     TaxIdRetrieveParams,
     TaxIdListParams,
     TaxIdDeleteParams,
+    TaxIdSerializeBatchCreateParams,
     TaxIdResource,
   };
   export {
@@ -2401,6 +2627,8 @@ export declare namespace Stripe {
     TaxRateRetrieveParams,
     TaxRateUpdateParams,
     TaxRateListParams,
+    TaxRateSerializeBatchCreateParams,
+    TaxRateSerializeBatchUpdateParams,
     TaxRateResource,
   };
   export {Token, TokenCreateParams, TokenRetrieveParams, TokenResource};
@@ -2443,11 +2671,16 @@ export declare namespace Stripe {
   export {Discount, DeletedDiscount};
   export {FundingInstructions};
   export {LineItem};
+  export {PaymentMethodBalance};
+  export {Profile};
   export {QuoteLine};
   export {ReserveTransaction};
+  export {RiskSignals};
   export {SourceMandateNotification};
   export {SourceTransaction};
   export {TaxDeductedAtSource};
+  export {TransitBalance};
+  export {AccountSignals};
   export {Capability};
   export {LoginLink};
   export {Person, DeletedPerson};
@@ -2469,11 +2702,13 @@ export declare namespace Stripe {
   export {Capital};
   export {Checkout};
   export {Climate};
+  export {DelegatedCheckout};
   export {Entitlements};
   export {FinancialConnections};
   export {Forwarding};
   export {Identity};
   export {Issuing};
+  export {Orchestration};
   export {Privacy};
   export {ProductCatalog};
   export {Radar};
@@ -2502,6 +2737,7 @@ export declare namespace Stripe {
     ApplicationFeeRefundedEvent,
     BalanceAvailableEvent,
     BalanceSettingsUpdatedEvent,
+    BillingAlertRecoveredEvent,
     BillingAlertTriggeredEvent,
     BillingCreditBalanceTransactionCreatedEvent,
     BillingCreditGrantCreatedEvent,
@@ -2653,6 +2889,7 @@ export declare namespace Stripe {
     PaymentIntentAmountCapturableUpdatedEvent,
     PaymentIntentCanceledEvent,
     PaymentIntentCreatedEvent,
+    PaymentIntentExpiredEvent,
     PaymentIntentPartiallyFundedEvent,
     PaymentIntentPaymentFailedEvent,
     PaymentIntentProcessingEvent,
@@ -2836,8 +3073,204 @@ export declare namespace Stripe {
   export {StripeContext as StripeContextType};
   export {StripeRawError};
   export {UnhandledNotificationDetails};
-  export import ErrorType = _Error;
   export import Events = V2.Core.Events;
+
+  // ErrorTypeNamespaces: The beginning of the section generated from our OpenAPI spec
+  export namespace ErrorType {
+    export type StripeError = InstanceType<typeof _Error.StripeError>;
+    export type StripeCardError = InstanceType<typeof _Error.StripeCardError>;
+    export type StripeInvalidRequestError = InstanceType<
+      typeof _Error.StripeInvalidRequestError
+    >;
+    export type StripeAPIError = InstanceType<typeof _Error.StripeAPIError>;
+    export type StripeAuthenticationError = InstanceType<
+      typeof _Error.StripeAuthenticationError
+    >;
+    export type StripePermissionError = InstanceType<
+      typeof _Error.StripePermissionError
+    >;
+    export type StripeRateLimitError = InstanceType<
+      typeof _Error.StripeRateLimitError
+    >;
+    export type StripeConnectionError = InstanceType<
+      typeof _Error.StripeConnectionError
+    >;
+    export type StripeSignatureVerificationError = InstanceType<
+      typeof _Error.StripeSignatureVerificationError
+    >;
+    export type StripeIdempotencyError = InstanceType<
+      typeof _Error.StripeIdempotencyError
+    >;
+    export type StripeOAuthError = InstanceType<typeof _Error.StripeOAuthError>;
+    export type StripeInvalidGrantError = InstanceType<
+      typeof _Error.StripeInvalidGrantError
+    >;
+    export type StripeInvalidClientError = InstanceType<
+      typeof _Error.StripeInvalidClientError
+    >;
+    export type StripeOAuthInvalidRequestError = InstanceType<
+      typeof _Error.StripeOAuthInvalidRequestError
+    >;
+    export type StripeInvalidScopeError = InstanceType<
+      typeof _Error.StripeInvalidScopeError
+    >;
+    export type StripeUnsupportedGrantTypeError = InstanceType<
+      typeof _Error.StripeUnsupportedGrantTypeError
+    >;
+    export type StripeUnsupportedResponseTypeError = InstanceType<
+      typeof _Error.StripeUnsupportedResponseTypeError
+    >;
+    export type AlreadyCanceledError = InstanceType<
+      typeof _Error.AlreadyCanceledError
+    >;
+    export type AlreadyExistsError = InstanceType<
+      typeof _Error.AlreadyExistsError
+    >;
+    export type BlockedByStripeError = InstanceType<
+      typeof _Error.BlockedByStripeError
+    >;
+    export type CannotProceedError = InstanceType<
+      typeof _Error.CannotProceedError
+    >;
+    export type ControlledByAlternateResourceError = InstanceType<
+      typeof _Error.ControlledByAlternateResourceError
+    >;
+    export type ControlledByDashboardError = InstanceType<
+      typeof _Error.ControlledByDashboardError
+    >;
+    export type FeatureNotEnabledError = InstanceType<
+      typeof _Error.FeatureNotEnabledError
+    >;
+    export type FinancialAccountNotOpenError = InstanceType<
+      typeof _Error.FinancialAccountNotOpenError
+    >;
+    export type FxQuoteExpiredError = InstanceType<
+      typeof _Error.FxQuoteExpiredError
+    >;
+    export type InsufficientFundsError = InstanceType<
+      typeof _Error.InsufficientFundsError
+    >;
+    export type InvalidPaymentMethodError = InstanceType<
+      typeof _Error.InvalidPaymentMethodError
+    >;
+    export type InvalidPayoutMethodError = InstanceType<
+      typeof _Error.InvalidPayoutMethodError
+    >;
+    export type NonZeroBalanceError = InstanceType<
+      typeof _Error.NonZeroBalanceError
+    >;
+    export type NotCancelableError = InstanceType<
+      typeof _Error.NotCancelableError
+    >;
+    export type QuotaExceededError = InstanceType<
+      typeof _Error.QuotaExceededError
+    >;
+    export type RateLimitError = InstanceType<typeof _Error.RateLimitError>;
+    export type RecipientNotNotifiableError = InstanceType<
+      typeof _Error.RecipientNotNotifiableError
+    >;
+    export type TemporarySessionExpiredError = InstanceType<
+      typeof _Error.TemporarySessionExpiredError
+    >;
+  }
+  export namespace errors {
+    export type StripeError = InstanceType<typeof _Error.StripeError>;
+    export type StripeCardError = InstanceType<typeof _Error.StripeCardError>;
+    export type StripeInvalidRequestError = InstanceType<
+      typeof _Error.StripeInvalidRequestError
+    >;
+    export type StripeAPIError = InstanceType<typeof _Error.StripeAPIError>;
+    export type StripeAuthenticationError = InstanceType<
+      typeof _Error.StripeAuthenticationError
+    >;
+    export type StripePermissionError = InstanceType<
+      typeof _Error.StripePermissionError
+    >;
+    export type StripeRateLimitError = InstanceType<
+      typeof _Error.StripeRateLimitError
+    >;
+    export type StripeConnectionError = InstanceType<
+      typeof _Error.StripeConnectionError
+    >;
+    export type StripeSignatureVerificationError = InstanceType<
+      typeof _Error.StripeSignatureVerificationError
+    >;
+    export type StripeIdempotencyError = InstanceType<
+      typeof _Error.StripeIdempotencyError
+    >;
+    export type StripeOAuthError = InstanceType<typeof _Error.StripeOAuthError>;
+    export type StripeInvalidGrantError = InstanceType<
+      typeof _Error.StripeInvalidGrantError
+    >;
+    export type StripeInvalidClientError = InstanceType<
+      typeof _Error.StripeInvalidClientError
+    >;
+    export type StripeOAuthInvalidRequestError = InstanceType<
+      typeof _Error.StripeOAuthInvalidRequestError
+    >;
+    export type StripeInvalidScopeError = InstanceType<
+      typeof _Error.StripeInvalidScopeError
+    >;
+    export type StripeUnsupportedGrantTypeError = InstanceType<
+      typeof _Error.StripeUnsupportedGrantTypeError
+    >;
+    export type StripeUnsupportedResponseTypeError = InstanceType<
+      typeof _Error.StripeUnsupportedResponseTypeError
+    >;
+    export type AlreadyCanceledError = InstanceType<
+      typeof _Error.AlreadyCanceledError
+    >;
+    export type AlreadyExistsError = InstanceType<
+      typeof _Error.AlreadyExistsError
+    >;
+    export type BlockedByStripeError = InstanceType<
+      typeof _Error.BlockedByStripeError
+    >;
+    export type CannotProceedError = InstanceType<
+      typeof _Error.CannotProceedError
+    >;
+    export type ControlledByAlternateResourceError = InstanceType<
+      typeof _Error.ControlledByAlternateResourceError
+    >;
+    export type ControlledByDashboardError = InstanceType<
+      typeof _Error.ControlledByDashboardError
+    >;
+    export type FeatureNotEnabledError = InstanceType<
+      typeof _Error.FeatureNotEnabledError
+    >;
+    export type FinancialAccountNotOpenError = InstanceType<
+      typeof _Error.FinancialAccountNotOpenError
+    >;
+    export type FxQuoteExpiredError = InstanceType<
+      typeof _Error.FxQuoteExpiredError
+    >;
+    export type InsufficientFundsError = InstanceType<
+      typeof _Error.InsufficientFundsError
+    >;
+    export type InvalidPaymentMethodError = InstanceType<
+      typeof _Error.InvalidPaymentMethodError
+    >;
+    export type InvalidPayoutMethodError = InstanceType<
+      typeof _Error.InvalidPayoutMethodError
+    >;
+    export type NonZeroBalanceError = InstanceType<
+      typeof _Error.NonZeroBalanceError
+    >;
+    export type NotCancelableError = InstanceType<
+      typeof _Error.NotCancelableError
+    >;
+    export type QuotaExceededError = InstanceType<
+      typeof _Error.QuotaExceededError
+    >;
+    export type RateLimitError = InstanceType<typeof _Error.RateLimitError>;
+    export type RecipientNotNotifiableError = InstanceType<
+      typeof _Error.RecipientNotNotifiableError
+    >;
+    export type TemporarySessionExpiredError = InstanceType<
+      typeof _Error.TemporarySessionExpiredError
+    >;
+  }
+  // ErrorTypeNamespaces: The end of the section generated from our OpenAPI spec
 }
 
 Stripe.initialize(new NodePlatformFunctions());